feat: add API "get app version by id"

+++ das_backend/src/main/java/ch/sbb/backend/admin/infrastructure/settings/AppVersionRepository.java
@@ -3,6 +3,6 @@ package ch.sbb.backend.admin.infrastructure.settings;
 import ch.sbb.backend.admin.infrastructure.settings.model.AppVersionEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AppVersionRepository extends JpaRepository<AppVersionEntity, Long> {
+public interface AppVersionRepository extends JpaRepository<AppVersionEntity, Integer> {
 
 }
